requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azlyrics==1.3.2
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==2.1.1
cloudscraper==1.2.68
colorama==0.4.6
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
Js2Py==0.74
lxml==4.9.2
lyricsgenius==3.0.1
multidict==6.0.4
openai==0.26.4
PloudosAPI==0.0.9
pyjsparser==2.7.1
pyparsing==3.0.9
python-aternos==2.1.3
python-dotenv==0.21.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
websockets==10.4
yarl==1.8.2
